Bugfix: Delete button returns when clicking in editable project name button
When clicking edit, the delete button dissapears as intended, but as soon as the user clicks in the editable space to move the cursor, the onclick event of the project button fires again.  This reruns the selection process unnecessarily again, as the project is already selected.  More importantly this re-selection shows the delete button again, before the user has finished editing the project name.  This bug is now fixed.